[TS PBT] Clarify mapping and process supervision
